+++ db/migrate/20260626120000_add_last_tool_name_to_agent_contexts.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddLastToolNameToAgentContexts < ActiveRecord::Migration[8.0]
+  def change
+    add_column :agent_contexts, :last_tool_name, :string
+  end
+end